% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendinfo.R
\name{append_ec_sites}
\alias{append_ec_sites}
\title{Append columns of enzyme class and active site information to the pairwise_LFQ output}
\usage{
append_ec_sites(LFQ_table, quantitation_level)
}
\arguments{
\item{LFQ_table}{a dataframe saved as the pairwise_LFQ() output}

\item{quantitation_level}{a string, must be either "peptide" or "protein", must be match the pairwise_LFQ argument}
}
\description{
Append columns of enzyme class and active site information to the pairwise_LFQ output
}
