% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancedsetoperations.R
\name{elemental_subset}
\alias{elemental_subset}
\title{Filter a dataset; drop rows where the cell (containing delimiter parsed elements) of exam_col does not contain a vector of specified values}
\usage{
elemental_subset(dataset, exam_col, cell_must_contain, delimiter)
}
\description{
Filter a dataset; drop rows where the cell (containing delimiter parsed elements) of exam_col does not contain a vector of specified values
}
