% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iodata.R
\name{make_custom_setList}
\alias{make_custom_setList}
\title{Function make_custom_setList}
\usage{
make_custom_setList(datasets = read_proteinGroups(), columnName)
}
\arguments{
\item{columnName}{a string that specifies the name of the column, which user wants to keep for making the setList.}
}
\description{
Convert a list of datasets generated from \code{\link{read_proteinGroups}} to a setList of protein groups to be used for Venn Diagram \code{\link{Max_Venn}}()  or Tier \code{\link{make_tiers}}()  analysis
}
