% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iodata.R
\name{make_proteinNames_setList}
\alias{make_proteinNames_setList}
\title{Function make_proteinNames_setList}
\usage{
make_proteinNames_setList(datasets = read_proteinGroups())
}
\description{
Convert a list of datasets generated from \code{\link{read_proteinGroups}} to a setList of protein groups to be used for Venn Diagram \code{\link{Max_Venn}}()  or Tier \code{\link{make_tiers}}()  analysis
}
