% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiers.R
\name{make_tiers}
\alias{make_tiers}
\title{Function make_tiers}
\usage{
make_tiers(setLists)
}
\description{
Generate a tier list from a set of lists of protein groups identified by MaxQuant for cross-list comparisons
}
