% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tieredsetintersection.R
\name{make_tiers2}
\alias{make_tiers2}
\title{Function make_tiers2}
\usage{
make_tiers2(setLists)
}
\arguments{
\item{setLists}{a setList of protein groups generated by \code{\link{make_proteinGroups_setList}}()}
}
\description{
Generate a tier list from a set of lists of protein groups identified by MaxQuant for cross-list comparisons. This function calls Max_Venn() and calculates sums of conventional Venn table generated by Max_Venn. Each protein group is treated as a single string for comparisons. Also see \code{\link{Max_Venn}}(IndividualAnalysis = FALSE)
}
