% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvenn.R
\name{Max_Venn}
\alias{Max_Venn}
\title{Function Max_Venn}
\usage{
Max_Venn(Sets, Weight, SetNames, numberOfSets, IndividualAnalysis)
}
\arguments{
\item{Sets}{a setList of protein groups generated by function \code{\link{make_proteinGroups_setList}}()}

\item{IndividualAnalysis}{a boolean. If FALSE, conventional intersection analysis is performed; each protein group is treated as a single string from comparisons. If TRUE, individual proteins in each protein group are compared to others; each protein group is treated as a vector of individual string elements for comparisons.}
}
\value{
an S4 class "Venn" object
}
\description{
Function Max_Venn
}
