% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplots.R
\name{multi_volcano_plots}
\alias{multi_volcano_plots}
\title{multi_volcano_plots: directly converts MaxQuant output to multiple volcano plots}
\usage{
multi_volcano_plots(
  raw = read.delim("modificationSpecificPeptides.txt", header = TRUE, sep = "\\t"),
  meta = read.delim("metadata.txt", header = TRUE, sep = "\\t"),
  name_probe_mod = c("Mod"),
  max_each_mod = 1,
  max_total_mods = 1,
  quantitation_level = "peptide",
  background_check = FALSE,
  normalize_to = NULL,
  xlim = c(-10, 3),
  ylim = c(0, 5),
  label_col_name = "Gene.Names",
  pCutoff = 0.05,
  FCcutoff = -2
)
}
\arguments{
\item{raw}{a dataframe by reading modificationSpecificPeptides.txt}

\item{name_probe_mod}{a string vector of chemical probe/modification names, such as c("Mod1", "Mod2"), must match MaxQuant input}

\item{max_each_mod}{a integer as the maximal number of modifications on a single peptide, set for each chemical probe}

\item{max_total_mods}{a integer as the maximal number of modifications on a single peptide, set for all chemical probes Note max_each_mod must not be less than max_total_mods}

\item{quantitation_level}{a string, must be either "peptide" or "protein"}

\item{background_check}{a boolean, FALSE = quantify probe-modified peptides, TRUE = quantify non-probe-modified peptides}

\item{normalize_to}{a string, must be either "sum_all", "mean_all", (normalize to all peptides) "sum_background", or "mean_background" (normalize to background/non-probe-modified peptides).}

\item{xlim}{a integer vector, such as c(-5, 5) for an x axis range of -5 to 5}

\item{ylim}{a integer vector, such as c(0, 5) for an y axis range of 0 to 5}

\item{label_col_name}{the input column name for labeling volcano plot data points such as "Gene.Names"}

\item{pCutoff}{the p-Value cutoff, for instance, default p-value = 0.05}

\item{FCcutoff}{the fold change cutoff, Note for ABPP, we are only interested in negative fold change (Lower intensity at higher inhibitor concentration)}

\item{metadata}{a dataframe that maches the MaxQuant input. Column 1: Intensity (such as Intensity samplename, same as the column names in modificationSpecificPeptides.txt) name Column 2: Replicate group (use the same name for each group of replicates)}
}
\value{
volcano plots
}
\description{
multi_volcano_plots: directly converts MaxQuant output to multiple volcano plots
}
\examples{
 multi_volcano_plots(raw = raw, meta = meta, name_probe_mod = c("Mod"),
                   max_each_mod = 1, max_total_mods = 1, quantitation_level = "peptide" , background_check = FALSE,
                   xlim = c(-10, 3), ylim = c(0, 5), label_col_name = "Gene.Names", pCutoff = 0.05, FCcutoff = -2)
}
\seealso{
\code{\link{pairwise_LFQ}}
         \code{\link{append_ec_sites}}
         \code{\link{plot_volcano}}
}
