% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxvenn.R
\name{plot_Max_Venn}
\alias{plot_Max_Venn}
\title{Function plot_Max_Venn}
\usage{
plot_Max_Venn(
  V = Max_Venn("setList"),
  doWeights = TRUE,
  add = FALSE,
  show = list(FaceText = "weight", Faces = TRUE),
  gpList
)
}
\arguments{
\item{V}{an S4 class "Venn" object generated by \code{\link{Max_Venn}}()}

\item{doWeights}{a boolean. If TRUE, plots weighed Venn Diagram (shapes change according to ID numbers in fields). If FALSE, plots unweighed Venn Diagram.}
}
\description{
Function plot_Max_Venn
}
