% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tieredsetintersection.R
\name{plot_target}
\alias{plot_target}
\title{Subroutine plot_target}
\usage{
plot_target(
  tiers,
  edges = 2000,
  density = 2000,
  angle = 45,
  border = FALSE,
  lty = NULL,
  title = NULL
)
}
\arguments{
\item{tiers}{a list generated by \code{\link{make_tiers}}() function.}

\item{title}{a string shown as the title. Default value is NULL.}
}
\description{
The plot_target Subroutine draws a target plot showing protein group ID numbers in Tiers.
}
