% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generates a volcano plot from the output of append_ec_sites()}
\usage{
plot_volcano(
  LFQ_table_ec,
  x,
  y,
  xlim,
  ylim,
  label_col_name,
  pCutoff,
  FCcutoff,
  title
)
}
\arguments{
\item{LFQ_table_ec}{a dataframe saved as the append_ec_sites() output}

\item{x}{log2FC column name}

\item{y}{-log10pvalue column name}

\item{xlim}{a integer vector, such as c(-5, 5) for an x axis range of -5 to 5}

\item{ylim}{a integer vector, such as c(0, 5) for an y axis range of 0 to 5}

\item{label_col_name}{the input column name for labeling volcano plot data points such as "Gene.Names"}

\item{pCutoff}{the p-Value cutoff on -log10 scale, for instance, use 1.3 for p-value = 0.05}

\item{FCcutoff}{the fold change cutoff on log2 scale, for instance, use -1 for -2 fold change, Note for ABPP, we are only interested in negative fold change}

\item{title}{a string as the title of the volcano plot}
}
\description{
Generates a volcano plot from the output of append_ec_sites()
}
