% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancedsetoperations.R
\name{proc_mspTable}
\alias{proc_mspTable}
\title{proc_mspTable: convert MaxQuant modificationSpecificPeptides.txt file to an intensity table}
\usage{
proc_mspTable(
  raw = read_tsv("modificationSpecificPeptides.txt"),
  metadata = read_tsv("metadata.txt"),
  name_probe_mod,
  max_each_mod = 1,
  max_total_mods = 1,
  quantitation_level = "peptide",
  background_check = FALSE,
  normalize_to = NULL
)
}
\arguments{
\item{raw}{a dataframe by reading modificationSpecificPeptides.txt}

\item{metadata}{a dataframe that maches the MaxQuant input. Column 1: Intensity (such as Intensity samplename, same as the column names in modificationSpecificPeptides.txt) name Column 2: Replicate group (use the same name for each group of replicates)}

\item{name_probe_mod}{a string vector of chemical probe/modification names, such as c("Mod1", "Mod2"), must match MaxQuant input}

\item{max_each_mod}{a integer as the maximal number of modifications on a single peptide, set for each chemical probe}

\item{max_total_mods}{a integer as the maximal number of modifications on a single peptide, set for all chemical probes Note max_each_mod must not be less than max_total_mods}

\item{quantitation_level}{a string, must be either "peptide" or "protein"}

\item{background_check}{a boolean, FALSE = quantify probe-modified peptides, TRUE = quantify non-probe-modified peptides}

\item{normalize_to}{a string, must be either "sum_all", "mean_all", (normalize to all peptides) "sum_background", or "mean_background" (normalize to background/non-probe-modified peptides).}
}
\value{
a tibble of extracted and process subset from modificationSpecificPeptides.txt
}
\description{
proc_mspTable: convert MaxQuant modificationSpecificPeptides.txt file to an intensity table
}
