% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iodata.R
\name{read_proteinGroups}
\alias{read_proteinGroups}
\title{Function read_proteinGroups}
\usage{
read_proteinGroups(
  folderName = "proteinGroups",
  selectColumnNames = defaultColumnNames
)
}
\arguments{
\item{folderName}{Default folder name is "proteinGroups", the fold must exist in working directory, use \code{\link{getwd}}() and \code{\link{dir}}() to check}

\item{selectColumnNames}{Default selected column names are
"Protein IDs",                          "Majority protein IDs",                 "Peptide counts (all)",
"Peptide counts (razor+unique)",        "Peptide counts (unique)",              "Protein names",
"Gene names",      "Number of proteins", "Peptides",                             "Razor + unique peptides",        "Unique peptides",
"Sequence coverage [%]",                "Unique + razor sequence coverage [%]", "Unique sequence coverage [%]",
"Mol. weight [kDa]",                    "Sequence length",                      "Sequence lengths",
"Q-value",                              "Score",                                    "Intensity",               "MS/MS count",
"Only identified by site",              "Potential contaminant"}
}
\value{
a list of dataframes (each dataframe contains data from each file)
}
\description{
Read all .txt files (proteinGroups.txt renamed according to a specific experimental condition) a user-created folder.
}
