% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancedsetoperations.R
\name{tibble_add_column}
\alias{tibble_add_column}
\title{Add a column of data to an empty tibble}
\usage{
tibble_add_column(data, ..., before = NULL, after = NULL)
}
\description{
Add a column of data to an empty tibble
}
