% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendinfo.R
\name{append_ec_sites}
\alias{append_ec_sites}
\title{Append columns of enzyme class and active site information to the pairwise_LFQ output}
\usage{
append_ec_sites(LFQ_table, quantitation_level)
}
\arguments{
\item{LFQ_table}{a dataframe saved as the pairwise_LFQ() output}

\item{quantitation_level}{a string, must be either "peptide" or "protein", must be match the pairwise_LFQ argument}
}
\value{
a volcano plot-ready dataframe annotated with two extra columns of enzyme class and active sites
}
\description{
Append columns of enzyme class and active site information to the pairwise_LFQ output
}
\examples{
output2 <- append_ec_sites(output1, quantitation_level = "peptide")
}
\seealso{
\code{\link{pairwise_LFQ}}
}
