% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiers.R
\name{make_tiers}
\alias{make_tiers}
\title{Function make_tiers}
\usage{
make_tiers(setLists)
}
\arguments{
\item{setLists}{a setList of protein groups generated by \code{\link{make_proteinGroups_setList}}()}
}
\description{
Generate a tier list from a set of lists of protein groups identified by MaxQuant for cross-list comparisons. This function calls Max_Venn() and calculates sums of redundant Venn table generated by Max_Venn. Individual proteins in each protein group are compared to others; each protein group is treated as a vector of individual string elements for comparisons. Also see \code{\link{Max_Venn}}(IndividualAnalysis = TRUE)
}
